Fix sequence.toml tab and space mix (1/2)
This toml file mixes tab and spaces, and displays properly when using 4
space tabs.

First, add a modeline that instruct vim to use the right number of
spaces to display the content of the file.  Alignment is still broken
when looking at the file through cat(1) / more(1), but is displayed
correctly in an editor with a default configuration.

While here, add a \n at the end of the file.